+++ mobile/android/android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/manager/State.kt
@@ -6,7 +6,6 @@ package mozilla.components.service.fxa.manager
 
 import mozilla.components.concept.sync.AuthType
 import mozilla.components.service.fxa.FxaAuthData
-import mozilla.components.service.fxa.sharing.ShareableAccount
 
 /**
  * This file is the "heart" of the accounts system. It's a finite state machine.
@@ -54,7 +53,6 @@ import mozilla.components.service.fxa.sharing.ShareableAccount
 
 internal enum class AccountState {
     NotAuthenticated,
-    IncompleteMigration,
     Authenticated,
     AuthenticationProblem,
 }
@@ -63,7 +61,6 @@ internal enum class ProgressState {
     Initializing,
     BeginningAuthentication,
     CompletingAuthentication,
-    MigratingAccount,
     RecoveringFromAuthProblem,
     LoggingOut,
 }
@@ -80,26 +77,15 @@ internal sealed class Event {
         }
         object AccessTokenKeyError : Account()
 
-        data class MigrateFromAccount(val account: ShareableAccount, val reuseSessionToken: Boolean) : Account() {
-            override fun toString(): String {
-                return this.javaClass.simpleName
-            }
-        }
-
-        object RetryMigration : Account()
-
         object Logout : Account()
     }
 
     internal sealed class Progress : Event() {
         object AccountNotFound : Progress()
         object AccountRestored : Progress()
-        data class IncompleteMigration(val reuseSessionToken: Boolean) : Progress()
 
         data class AuthData(val authData: FxaAuthData) : Progress()
-        data class Migrated(val reusedSessionToken: Boolean) : Progress()
 
-        object FailedToCompleteMigration : Progress()
         object FailedToBeginAuth : Progress()
         object FailedToCompleteAuthRestore : Progress()
         object FailedToCompleteAuth : Progress()
@@ -127,12 +113,6 @@ internal fun State.next(event: Event): State? = when (this) {
             Event.Account.Start -> State.Active(ProgressState.Initializing)
             Event.Account.BeginEmailFlow -> State.Active(ProgressState.BeginningAuthentication)
             is Event.Account.BeginPairingFlow -> State.Active(ProgressState.BeginningAuthentication)
-            is Event.Account.MigrateFromAccount -> State.Active(ProgressState.MigratingAccount)
-            else -> null
-        }
-        AccountState.IncompleteMigration -> when (event) {
-            is Event.Account.RetryMigration -> State.Active(ProgressState.MigratingAccount)
-            is Event.Account.Logout -> State.Active(ProgressState.LoggingOut)
             else -> null
         }
         AccountState.Authenticated -> when (event) {
@@ -152,7 +132,6 @@ internal fun State.next(event: Event): State? = when (this) {
         ProgressState.Initializing -> when (event) {
             Event.Progress.AccountNotFound -> State.Idle(AccountState.NotAuthenticated)
             Event.Progress.AccountRestored -> State.Active(ProgressState.CompletingAuthentication)
-            is Event.Progress.IncompleteMigration -> State.Active(ProgressState.MigratingAccount)
             else -> null
         }
         ProgressState.BeginningAuthentication -> when (event) {
@@ -166,12 +145,6 @@ internal fun State.next(event: Event): State? = when (this) {
             is Event.Account.AuthenticationError -> State.Active(ProgressState.RecoveringFromAuthProblem)
             Event.Progress.FailedToCompleteAuthRestore -> State.Idle(AccountState.NotAuthenticated)
             Event.Progress.FailedToCompleteAuth -> State.Idle(AccountState.NotAuthenticated)
-            else -> null
-        }
-        ProgressState.MigratingAccount -> when (event) {
-            is Event.Progress.Migrated -> State.Active(ProgressState.CompletingAuthentication)
-            Event.Progress.FailedToCompleteMigration -> State.Idle(AccountState.NotAuthenticated)
-            is Event.Progress.IncompleteMigration -> State.Idle(AccountState.IncompleteMigration)
             else -> null
         }
         ProgressState.RecoveringFromAuthProblem -> when (event) {