+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/home/ModeTest.kt
@@ -79,7 +79,6 @@ class ModeTest {
     fun `get current onboarding mode when signed out`() {
         every { onboarding.userHasBeenOnboarded() } returns false
         every { accountManager.authenticatedAccount() } returns null
-        every { accountManager.shareableAccounts(context) } returns emptyList()
 
         assertEquals(Mode.Onboarding(OnboardingState.SignedOutNoAutoSignIn, onboarding.config), currentMode.getCurrentMode())
     }