+++ mobile/android/docs/changelog.md
@@ -10,8 +10,11 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/firefox-android/blob/main/android-components/plugins/dependencies/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/firefox-android/blob/main/android-components/.config.yml)
 
-* * **compose-cfr**
-* 🚒 Bug fixed [bug #1819950](https://bugzilla.mozilla.org/show_bug.cgi?id=1819950). Ensure CFRs are automatically dismissed on screen rotation on all Android versions.
+* **compose-cfr**
+  * 🚒 Bug fixed [bug #1819950](https://bugzilla.mozilla.org/show_bug.cgi?id=1819950). Ensure CFRs are automatically dismissed on screen rotation on all Android versions.
+
+* **concept-sync**, **service-firefox-accounts**
+  * Removed unused `AccountSharing` from sync and accounts.
 
 # 112.0.0
 * [Commits](https://github.com/mozilla-mobile/firefox-android/compare/releases_v111...releases_v112)